use a dedicated Dockerfile for the scenario test(s)
The initial Dockerfile is optimized for running during a Dockerhub
build. If module cache is empty the whole test gets slow because the
image build depends on the Go build which depends on module loading.

In most of the cases a 'go build' has happened before 'go test'. Thus,
the fakedata binary should already be present and can be used for
building the Docker image.